+++ packages/url_launcher/url_launcher/android/src/main/java/io/flutter/plugins/urllauncher/UrlLauncher.java
@@ -1,4 +1,4 @@
-package dev.flutter.plugins.urllauncher;
+package io.flutter.plugins.urllauncher;
 
 import android.app.Activity;
 import android.content.ComponentName;
