+++ packages/url_launcher/url_launcher/android/src/main/java/io/flutter/plugins/urllauncher/UrlLauncherPluginRegistrar.java
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package dev.flutter.plugins.urllauncher;
+package io.flutter.plugins.urllauncher;
 
 import io.flutter.plugin.common.PluginRegistry.Registrar;
