+++ packages/url_launcher/url_launcher/android/src/test/java/io/flutter/plugins/urllauncher/MethodCallHandlerImplTest.java
@@ -1,4 +1,4 @@
-package dev.flutter.plugins.urllauncher;
+package io.flutter.plugins.urllauncher;
 
 import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.eq;
