+++ packages/url_launcher/url_launcher/pubspec.yaml
@@ -7,7 +7,7 @@ version: 5.1.7
 
 flutter:
   plugin:
-    androidPackage: dev.flutter.plugins.urllauncher
+    androidPackage: io.flutter.plugins.urllauncher
     iosPrefix: FLT
     pluginClass: UrlLauncherPluginRegistrar
